+++ graphbrain/parsers/parser_de.pyx
@@ -1,35 +1,35 @@
 import logging
-import spacy
 from graphbrain import *
 from .alpha_beta import AlphaBeta
 from .nlp import token2str
 
 
 deps_arg_roles = {
-    'nsubj': 's',      # subject
-    'nsubjpass': 'p',  # passive subject
-    'agent': 'a',      # agent
-    'acomp': 'c',      # subject complement
-    'attr': 'c',       # subject complement
-    'dobj': 'o',       # direct object
-    'prt': 'o',        # direct object
-    'dative': 'i',     # indirect object
-    'advcl': 'x',      # specifier
-    'prep': 'x',       # specifier
-    'npadvmod': 'x',   # specifier
-    'parataxis': 't',  # parataxis
-    'intj': 'j',       # interjection
-    'xcomp': 'r',      # clausal complement
-    'ccomp': 'r'       # clausal complement
+    'sb': 's',      # subject
+    'pd': 'c',      # subject complement
+    # 'nsubjpass': 'p',  # passive subject
+    # 'agent': 'a',      # agent
+    # 'acomp': 'c',      # subject complement
+    # 'attr': 'c',       # subject complement
+    # 'dobj': 'o',       # direct object
+    # 'prt': 'o',        # direct object
+    # 'dative': 'i',     # indirect object
+    # 'advcl': 'x',      # specifier
+    # 'prep': 'x',       # specifier
+    # 'npadvmod': 'x',   # specifier
+    # 'parataxis': 't',  # parataxis
+    # 'intj': 'j',       # interjection
+    # 'xcomp': 'r',      # clausal complement
+    # 'ccomp': 'r'       # clausal complement
 }
 
 
-# DONE
+##
 def is_noun(token):
     return token.tag_[0] == 'N'
 
 
-# DONE
+##
 def is_verb(token):
     tag = token.tag_
     if len(tag) > 0:
@@ -38,29 +38,27 @@ def is_verb(token):
         return False
 
 
-# DONE
+##
 def is_infinitive(token):
     return token.tag_ in {'VVINF', 'VVIZU'}
 
-
+##
 def concept_type_and_subtype(token):
     tag = token.tag_
     dep = token.dep_
-    if dep == 'nmod':
-        return 'cm'
-    if tag[:2] == 'JJ':
+    if tag[:3] == 'ADJ':
         return 'ca'
-    elif tag[:2] == 'NN':
-        subtype = 'p' if 'P' in tag else 'c'
-        sing_plur = 'p' if tag[-1] == 'S' else 's'
-        return 'c{}.{}'.format(subtype, sing_plur)
-    elif tag == 'CD':
+    elif tag == 'NN':
+        return 'cc'
+    elif tag in {'NE', 'NNE'}:
+        return 'cp'
+    elif tag == 'CARD':
         return 'c#'
-    elif tag == 'DT':
-        return 'cd'
-    elif tag == 'WP':
-        return 'cw'
-    elif tag == 'PRP':
+    # elif tag == 'DT':
+    #     return 'cd'
+    # elif tag == 'WP':
+    #     return 'cw'
+    elif tag == 'PPER':
         return 'ci'
     else:
         return 'c'
@@ -68,30 +66,23 @@ def concept_type_and_subtype(token):
 
 def modifier_type_and_subtype(token):
     tag = token.tag_
-    if tag == 'JJ':
-        return 'ma'
-    elif tag == 'JJR':
-        return 'mc'
-    elif tag == 'JJS':
-        return 'ms'
-    elif tag == 'DT':
+    if tag == 'ART':
         return 'md'
-    # elif tag == 'PDT':
-    #     return 'mp'
-    elif tag == 'WDT':
-        return 'mw'
-    elif tag == 'CD':
-        return 'm#'
     else:
         return 'm'
 
 
 def builder_type_and_subtype(token):
+    if token.head:
+        if token.head.dep_ == 'ag':
+            return 'bp'
+        
     tag = token.tag_
-    if tag == 'IN':
-        return 'br'  # relational (proposition)
-    elif tag == 'CC':
+    if tag == 'KON':
         return 'b+'
+    # still english below...
+    elif tag == 'IN':
+        return 'br'  # relational (proposition)
     elif tag == 'DT':
         return 'bd'
     else:
@@ -102,6 +93,7 @@ def arg_type(token):
     return deps_arg_roles.get(token.dep_, '?')
 
 
+# done
 def _verb_features(token):
     verb_form = '-'
     tense = '-'
@@ -109,30 +101,45 @@ def _verb_features(token):
     mood = '-'
     person = '-'
     number = '-'
+    verb_type = '-'
 
-    if token.tag_ == 'VB':
+    if token.tag_ == 'VAFIN':
+        mood = 'i'  # indicative
+        verb_form = 'f'  # finite
+    elif token.tag_ == 'VAIMP':
+        mood = '!'  # imperative
+        verb_form = 'f'  # finite
+    elif token.tag_ == 'VAINF':
         verb_form = 'i'  # infinitive
-    elif token.tag_ == 'VBD':
-        verb_form = 'f'  # finite
-        tense = '<'  # past
-    elif token.tag_ == 'VBG':
-        verb_form = 'p'  # participle
-        tense = '|'  # present
-        aspect = 'g'  # progressive
-    elif token.tag_ == 'VBN':
-        verb_form = 'p'  # participle
-        tense = '<'  # past
+    elif token.tag_ == 'VAPP':
         aspect = 'f'  # perfect
-    elif token.tag_ == 'VBP':
         verb_form = 'f'  # finite
-        tense = '|'  # present
-    elif token.tag_ == 'VBZ':
+    elif token.tag_ == 'VMFIN':
+        mood = 'i'  # indicative
         verb_form = 'f'  # finite
-        tense = '|'  # present
-        number = 's'  # singular
-        person = '3'  # third person
+        verb_type = 'm'  # modal
+    elif token.tag_ == 'VMINF':
+        verb_form = 'f'  # finite
+        verb_type = 'm'  # modal
+    elif token.tag_ == 'VMPP':
+        aspect = 'f'  # perfect
+        verb_form = 'p'  # participle
+        verb_type = 'm'  # modal
+    elif token.tag_ == 'VVFIN':
+        mood = 'i'  # indicative
+        verb_form = 'f'  # finite
+    elif token.tag_ == 'VVIMP':
+        mood = '!'  # imperative
+        verb_form = 'f'  # finite
+    elif token.tag_ == 'VVINF':
+        verb_form = 'i'  # infinitive
+    elif token.tag_ == 'VVIZU':
+        verb_form = 'i'  # infinitive
+    elif token.tag_ == 'VVPP':
+        aspect = 'f'  # perfect
+        verb_form = 'p'  # participle
 
-    features = (tense, verb_form, aspect, mood, person, number)
+    features = (tense, verb_form, aspect, mood, person, number, verb_type)
     return ''.join(features)
 
 
@@ -140,7 +147,7 @@ class ParserDE(AlphaBeta):
     def __init__(self, lemmas=False):
         super().__init__(lemmas=lemmas)
         self.lang = 'de'
-        self.nlp = spacy.load('de_core_news_md')
+        self.nlp = nlp
 
     # ===========================================
     # Implementation of language-specific methods
@@ -159,6 +166,18 @@ class ParserDE(AlphaBeta):
                     return 'm'
             return 'a'
 
+    def _builder_arg_roles(self, edge):
+        connector = edge[0]
+        if connector.is_atom():
+            ct = connector.type()
+            if ct == 'br':
+                connector = connector.replace_atom_part(
+                    1, '{}.ma'.format(ct))
+            elif ct == 'bp':
+                connector = connector.replace_atom_part(
+                    1, '{}.ma'.format(ct))
+        return hedge((connector,) + edge[1:])
+        
     def _build_atom(self, token, ent_type, ps):
         text = token.text.lower()
         et = ent_type
@@ -178,7 +197,7 @@ class ParserDE(AlphaBeta):
     def _token_type(self, token, head=False):
         dep = token.dep_
 
-        if dep in {'', 'subtok'}:
+        if dep == 'punct':
             return None
 
         head_type = self._token_head_type(token)
@@ -189,95 +208,24 @@ class ParserDE(AlphaBeta):
         if len(head_type) > 0:
             head_type = head_type[0]
 
-        if dep == 'ROOT':
-            if is_verb(token):
-                return 'p'
-            else:
-                return concept_type_and_subtype(token)
-        elif dep in {'appos', 'attr', 'compound', 'dative', 'dep', 'dobj',
-                     'nsubj', 'nsubjpass', 'oprd', 'pobj', 'meta'}:
+        if is_noun(token):
             return concept_type_and_subtype(token)
-        elif dep in {'advcl', 'csubj', 'csubjpass', 'parataxis'}:
-            return 'p'
-        elif dep in {'relcl', 'ccomp'}:
+        elif dep == 'ROOT':
             if is_verb(token):
-                return 'pr'
-            else:
-                return concept_type_and_subtype(token)
-        elif dep in {'acl', 'pcomp', 'xcomp'}:
-            if token.tag_ == 'IN':
-                return 'a'
-            else:
-                return 'pc'
-        elif dep in {'amod', 'nummod', 'preconj', 'predet'}:
-            return modifier_type_and_subtype(token)
-        elif dep == 'det':
-            if token.head.dep_ == 'npadvmod':
-                return builder_type_and_subtype(token)
-            else:
-                return modifier_type_and_subtype(token)
-        elif dep in {'aux', 'auxpass', 'expl', 'prt', 'quantmod'}:
-            if head_type == 'c':
-                return 'm'
-            if token.n_lefts + token.n_rights == 0:
-                return 'a'
-            else:
-                return 'x'
-        elif dep in {'nmod', 'npadvmod'}:
-            if is_noun(token):
-                return concept_type_and_subtype(token)
-            else:
-                return modifier_type_and_subtype(token)
-        elif dep == 'cc':
-            if head_type == 'p':
-                return 'pm'
-            else:
-                return builder_type_and_subtype(token)
-        elif dep == 'case':
-            if token.head.dep_ == 'poss':
-                return 'bp'
-            else:
-                return builder_type_and_subtype(token)
-        elif dep == 'neg':
-            return 'an'
-        elif dep == 'agent':
-            return 'x'
-        elif dep in {'intj', 'punct'}:
-            return ''
-        elif dep == 'advmod':
-            if token.head.dep_ == 'advcl':
-                return 't'
-            elif head_type == 'p':
-                return 'a'
-            elif head_type in {'m', 'x', 't', 'b'}:
-                return 'w'
-            else:
-                return modifier_type_and_subtype(token)
-        elif dep == 'poss':
-            if is_noun(token):
-                return concept_type_and_subtype(token)
-            else:
-                return 'mp'
-        elif dep == 'prep':
-            if head_type == 'p':
-                return 't'
-            else:
-                return builder_type_and_subtype(token)
-        elif dep == 'conj':
-            if head_type == 'p' and is_verb(token):
                 return 'p'
             else:
                 return concept_type_and_subtype(token)
-        elif dep == 'mark':
-            if head_type == 'p' and head_subtype != 'c':
-                return 'x'
-            else:
+        elif dep in {'sb', 'pd', 'ag'}:
+            return concept_type_and_subtype(token)
+        elif dep in {'cj', 'mo'}:
+            return modifier_type_and_subtype(token)
+        elif dep == 'nk':
+            if token.head.dep_ == 'ag':
                 return builder_type_and_subtype(token)
-        elif dep == 'acomp':
-            if is_verb(token):
-                return 'x'
             else:
-                return concept_type_and_subtype(token)
+                return modifier_type_and_subtype(token)
+        elif dep == 'cd':
+            return builder_type_and_subtype(token)
         else:
             logging.warning('Unknown dependency (token_type): token: {}'
                             .format(token2str(token)))
@@ -293,6 +241,7 @@ class ParserDE(AlphaBeta):
     # Private methods
     # ===============
 
+    # move this to parent class
     def _token_head_type(self, token):
         head = token.head
         if head and head != token: